chore: Refactor
* Introduce feature grouping, collecting components,
  actions, reducers and tests per feature.

solves: #25
