+++ src/components/TabIcon.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
-import { colors } from './constants';
+import { colors } from '../api/constants';
 
 
 export const TabIcon = (icon) => {
