+++ src/features/Auth/actions/index.js
@@ -1,2 +1,2 @@
-export * from './AuthActions';
+export * from './LoginActions';
 export * from './RegisterActions';
