+++ src/features/News/components/NewsDisplay.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Text } from 'react-native';
 import { connect } from 'react-redux';
-import { Card, CardSection } from './common';
+import { Card, CardSection } from '../../../components';
 
 export class NewsDisplay extends Component {
