+++ src/features/Songs/actions/types.js
@@ -1,0 +1,4 @@
+export const SONG_UPDATE = 'song_update';
+export const SONG_CREATE = 'song_create';
+export const SONGS_FETCH_SUCCESS = 'song_fetch_success';
+export const SONG_SAVE_SUCCESS = 'song_save_success';
