--- src/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import AuthReducer from './AuthReducer';
-import RegisterReducer from './RegisterReducer';
-
-export default combineReducers({
-  auth: AuthReducer,
-  register: RegisterReducer,
-});
